Add initial UWyo sounding support
Add support for plotting Skew-T from uwyo sounding data.  This is
preliminary and real data is messy so there are bound to be some issues
that pop up.  Two testcases are present in the `testdata/` directory and
any other soundings will need to be in the same format (title line,
blank line, 4 lines of header, then data).  To plot these files, use the
`skewt` command line script (see `skewt --help` for details).